// utils/graphql/storefront.graphql.gen.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
// ...
export type Scalars = {
  ID: { input: string; output: string };
  String: { input: string; output: string };
  Boolean: { input: boolean; output: boolean };
  Int: { input: number; output: number };
  Float: { input: number; output: number };
  Any: { input: any; output: any };
  CEP: { input: any; output: any };
  CountryCode: { input: any; output: any };
  DateTime: { input: any; output: any };
  Decimal: { input: any; output: any };
  EmailAddress: { input: any; output: any };
  Long: { input: any; output: any };
  Upload: { input: any; output: any };
  Uuid: { input: any; output: any };
};
// ...
export type BuyBox = {
  /** List of the possibles installment plans. */
  installmentPlans?: Maybe<Array<Maybe<InstallmentPlan>>>;
  /** Maximum price among sellers. */
  maximumPrice?: Maybe<Scalars["Decimal"]["output"]>;
  /** Minimum price among sellers. */
  minimumPrice?: Maybe<Scalars["Decimal"]["output"]>;
  /** Quantity of offers. */
  quantityOffers?: Maybe<Scalars["Int"]["output"]>;
  /** List of sellers. */
  sellers?: Maybe<Array<Maybe<Seller>>>;
};
// ...
export type InstallmentPlan = {
  /** The custom display name of this installment plan. */
  displayName?: Maybe<Scalars["String"]["output"]>;
  /** List of the installments. */
  installments?: Maybe<Array<Maybe<Installment>>>;
  /** The name of this installment plan. */
  name?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type Operation =
  /** Performs AND operation between query terms. */
  | "AND"
  /** Performs OR operation between query terms. */
  | "OR";
// ...
export type PartnerSortKeys =
  /** The partner unique identifier. */
  | "ID"
  /** The partner name. */
  | "NAME";
// ...
export type ProductFilterInput = {
  /** The attribute name. */
  field: Scalars["String"]["input"];
  /** The set of values which the result filter item value must be included in. */
  values: Array<InputMaybe<Scalars["String"]["input"]>>;
};
// ...
export type ProductSearchSortKeys =
  /** The applied discount to the product variant price. */
  | "DISCOUNT"
  /** The product name. */
  | "NAME"
  /** The product variant price. */
  | "PRICE"
  /** Sort in a random way. */
  | "RANDOM"
  /** The date the product was released. */
  | "RELEASE_DATE"
  /** The relevance that the search engine gave to the possible result item based on own criteria. */
  | "RELEVANCE"
  /** The sales number on a period of time. */
  | "SALES"
  /** The quantity in stock of the product variant. */
  | "STOCK";

/** Define the product attribute which the result set will be sorted on. */
export type ProductSortKeys =
  /** The applied discount to the product variant price. */
  | "DISCOUNT"
  /** The product name. */
  | "NAME"
  /** The product variant price. */
  | "PRICE"
  /** Sort in a random way. */
  | "RANDOM"
  /** The date the product was released. */
  | "RELEASE_DATE"
  /** The sales number on a period of time. */
  | "SALES"
  /** The quantity in stock of the product variant. */
  | "STOCK";
// ...
export type Seller = {
  /** Seller name */
  name?: Maybe<Scalars["String"]["output"]>;
};
// ...
export type SortDirection =
  /** The results will be sorted in an ascending order. */
  | "ASC"
  /** The results will be sorted in an descending order. */
  | "DESC";
// ...
export type ProductFragment = {
  mainVariant?: boolean | null;
  productName?: string | null;
  productId?: any | null;
  alias?: string | null;
  available?: boolean | null;
  averageRating?: number | null;
  condition?: string | null;
  createdAt?: any | null;
  ean?: string | null;
  id?: string | null;
  minimumOrderQuantity?: number | null;
  productVariantId?: any | null;
  parentId?: any | null;
  sku?: string | null;
  numberOfVotes?: number | null;
  stock?: any | null;
  variantName?: string | null;
  variantStock?: any | null;
  collection?: string | null;
  urlVideo?: string | null;
  attributes?: Array<{
    value?: string | null;
    name?: string | null;
    attributeId?: string | null;
    displayType?: string | null;
    id?: string | null;
  } | null> | null;
  productCategories?: Array<{
    id: number;
    name?: string | null;
    url?: string | null;
    hierarchy?: string | null;
    main: boolean;
    googleCategories?: string | null;
  } | null> | null;
  informations?: Array<{
    title?: string | null;
    value?: string | null;
    type?: string | null;
  } | null> | null;
  images?: Array<{
    url?: string | null;
    fileName?: string | null;
    print: boolean;
  } | null> | null;
  prices?: {
    discountPercentage: any;
    discounted: boolean;
    listPrice?: any | null;
    multiplicationFactor: number;
    price: any;
    bestInstallment?: {
      discount: boolean;
      displayName?: string | null;
      fees: boolean;
      name?: string | null;
      number: number;
      value: any;
    } | null;
    installmentPlans?: Array<{
      displayName?: string | null;
      name?: string | null;
      installments?: Array<{
        discount: boolean;
        fees: boolean;
        number: number;
        value: any;
      } | null> | null;
    } | null> | null;
    priceTables?: Array<{
      discountPercentage: any;
      id: any;
      listPrice?: any | null;
      price: any;
    } | null> | null;
    wholesalePrices?: Array<{ price: any; quantity: number } | null> | null;
  } | null;
  productBrand?: {
    fullUrlLogo?: string | null;
    logoUrl?: string | null;
    name?: string | null;
    alias?: string | null;
  } | null;
  seller?: { name?: string | null } | null;
  similarProducts?: Array<{
    alias?: string | null;
    image?: string | null;
    imageUrl?: string | null;
    name?: string | null;
  } | null> | null;
  promotions?: Array<{
    content?: string | null;
    disclosureType?: string | null;
    id: any;
    fullStampUrl?: string | null;
    stamp?: string | null;
    title?: string | null;
  } | null> | null;
  buyBox?: BuyBox | null;
};
// ...
export type SearchQueryVariables = Exact<{
  operation: Operation;
  query?: InputMaybe<Scalars["String"]["input"]>;
  onlyMainVariant?: InputMaybe<Scalars["Boolean"]["input"]>;
  minimumPrice?: InputMaybe<Scalars["Decimal"]["input"]>;
  maximumPrice?: InputMaybe<Scalars["Decimal"]["input"]>;
  limit?: InputMaybe<Scalars["Int"]["input"]>;
  offset?: InputMaybe<Scalars["Int"]["input"]>;
  sortDirection?: InputMaybe<SortDirection>;
  sortKey?: InputMaybe<ProductSearchSortKeys>;
  filters?: InputMaybe<
    Array<InputMaybe<ProductFilterInput>> | InputMaybe<ProductFilterInput>
  >;
  partnerAccessToken?: InputMaybe<Scalars["String"]["input"]>;
  ignoreDisplayRules?: boolean;
}>;

export type SearchQuery = {
  result?: {
    pageSize: number;
    redirectUrl?: string | null;
    searchTime?: string | null;
    aggregations?: {
      maximumPrice: any;
      minimumPrice: any;
      priceRanges?: Array<{
        quantity: number;
        range?: string | null;
      } | null> | null;
      filters?: Array<{
        field?: string | null;
        origin?: string | null;
        values?: Array<{
          quantity: number;
          name?: string | null;
        } | null> | null;
      } | null> | null;
    } | null;
    breadcrumbs?: Array<{
      link?: string | null;
      text?: string | null;
    } | null> | null;
    forbiddenTerm?: { text?: string | null; suggested?: string | null } | null;
    productsByOffset?: {
      page: number;
      pageSize: number;
      totalCount: number;
      items?: Array<{
        mainVariant?: boolean | null;
        productName?: string | null;
        productId?: any | null;
        alias?: string | null;
        available?: boolean | null;
        averageRating?: number | null;
        condition?: string | null;
        createdAt?: any | null;
        ean?: string | null;
        id?: string | null;
        minimumOrderQuantity?: number | null;
        productVariantId?: any | null;
        parentId?: any | null;
        sku?: string | null;
        numberOfVotes?: number | null;
        stock?: any | null;
        variantName?: string | null;
        variantStock?: any | null;
        collection?: string | null;
        urlVideo?: string | null;
        attributes?: Array<{
          value?: string | null;
          name?: string | null;
        } | null> | null;
        productCategories?: Array<{
          id: number;
          name?: string | null;
          url?: string | null;
          hierarchy?: string | null;
          main: boolean;
          googleCategories?: string | null;
        } | null> | null;
        informations?: Array<{
          title?: string | null;
          value?: string | null;
          type?: string | null;
        } | null> | null;
        images?: Array<{
          url?: string | null;
          fileName?: string | null;
          print: boolean;
        } | null> | null;
        prices?: {
          discountPercentage: any;
          discounted: boolean;
          listPrice?: any | null;
          multiplicationFactor: number;
          price: any;
          bestInstallment?: {
            discount: boolean;
            displayName?: string | null;
            fees: boolean;
            name?: string | null;
            number: number;
            value: any;
          } | null;
          installmentPlans?: Array<{
            displayName?: string | null;
            name?: string | null;
            installments?: Array<{
              discount: boolean;
              fees: boolean;
              number: number;
              value: any;
            } | null> | null;
          } | null> | null;
          priceTables?: Array<{
            discountPercentage: any;
            id: any;
            listPrice?: any | null;
            price: any;
          } | null> | null;
          wholesalePrices?: Array<{
            price: any;
            quantity: number;
          } | null> | null;
        } | null;
        productBrand?: {
          fullUrlLogo?: string | null;
          logoUrl?: string | null;
          name?: string | null;
          alias?: string | null;
        } | null;
        seller?: { name?: string | null } | null;
        similarProducts?: Array<{
          alias?: string | null;
          image?: string | null;
          imageUrl?: string | null;
          name?: string | null;
        } | null> | null;
        promotions?: Array<{
          content?: string | null;
          disclosureType?: string | null;
          id: any;
          fullStampUrl?: string | null;
          stamp?: string | null;
          title?: string | null;
        } | null> | null;
        buyBox?: BuyBox | null;
        spotInformation?: Maybe<Scalars["String"]["output"]>;
      } | null> | null;
    } | null;
  } | null;
};
// ...
export type GetPartnersQueryVariables = Exact<{
  first?: InputMaybe<Scalars["Int"]["input"]>;
  last?: InputMaybe<Scalars["Int"]["input"]>;
  names?: InputMaybe<
    | Array<InputMaybe<Scalars["String"]["input"]>>
    | InputMaybe<Scalars["String"]["input"]>
  >;
  priceTableIds?: InputMaybe<
    Array<Scalars["Int"]["input"]> | Scalars["Int"]["input"]
  >;
  sortDirection?: SortDirection;
  sortKey?: PartnerSortKeys;
  before?: InputMaybe<Scalars["String"]["input"]>;
  alias?: InputMaybe<
    | Array<InputMaybe<Scalars["String"]["input"]>>
    | InputMaybe<Scalars["String"]["input"]>
  >;
  after?: InputMaybe<Scalars["String"]["input"]>;
}>;

export type GetPartnersQuery = {
  partners?: {
    edges?: Array<{
      node?: {
        partnerId: any;
        priceTableId: number;
        portfolioId: number;
        type?: string | null;
        startDate: any;
        endDate: any;
        name?: string | null;
        alias?: string | null;
        fullUrlLogo?: string | null;
        origin?: string | null;
        partnerAccessToken?: string | null;
      } | null;
    }> | null;
  } | null;
};